refactor: convert comment plugin helper functions into hook
